+++ changelog.md
@@ -67,9 +67,11 @@
   - Help:
     - Search correctly goes to first found result when text changed
     - Fix search scrolling offsetting main help window
-  - Backup: Fix importing of paths when not defined
+  - Backup: 
+    - Fix importing of paths when not defined
+    - Fix importing settings
   - Progress: Fix progress dialogs
-  - Mapper: fix export, remove, and other related buttons/menu items
+  - Mapper: Fix export, remove, and other related buttons/menu items
 - **Changed:**
   - `Character manager on row double click load in:` Now defaults to open in new tab
   - Character manger: